requirements.txt
streamlit
opencv-python
torch
numpy
ultralytics
matplotlib
python-dotenv
pillow
